% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_time_series_model.R
\name{as_model_list}
\alias{as_model_list}
\title{Coerce to a list of models}
\usage{
as_model_list(x)
}
\arguments{
\item{x}{A \code{time_series_model} or \code{sum_model}.}
}
\value{
List of \code{time_series_model} objects.
}
\description{
Helper used by \code{+} methods to normalize inputs.
}
\keyword{internal}
