% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwm2.R
\name{.comp_prefix}
\alias{.comp_prefix}
\title{Component name prefix for summed models}
\usage{
.comp_prefix(i)
}
\arguments{
\item{i}{Component index.}
}
\value{
Prefix string like \code{m1_}.
}
\description{
Component name prefix for summed models
}
\keyword{internal}
