% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{.gmwmx2_get_plot_colors}
\alias{.gmwmx2_get_plot_colors}
\title{Get plot colors for generated time series}
\usage{
.gmwmx2_get_plot_colors(n)
}
\arguments{
\item{n}{Number of colors to return.}
}
\value{
Character vector of hex colors.
}
\description{
Returns a vector of plot colors of length \code{n}. If \code{n} exceeds the size of
\code{gmwmx2_plot_colors}, a larger palette is created via interpolation.
}
\seealso{
gmwmx2_plot_colors
}
\keyword{internal}
