% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngl.R
\name{download_all_stations_names_ngl}
\alias{download_all_stations_names_ngl}
\title{Download all stations name from Nevada Geodetic Laboratory}
\usage{
download_all_stations_names_ngl()
}
\description{
Download all stations name from Nevada Geodetic Laboratory
}
