% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngl.R
\name{download_all_stations_ngl}
\alias{download_all_stations_ngl}
\title{Download all stations name and location from Nevada Geodetic Laboratory (NGL)}
\usage{
download_all_stations_ngl()
}
\value{
Return a \code{data.frame} with all stations name, latitude, longitude and heights.
}
\description{
Download all stations name and location from Nevada Geodetic Laboratory (NGL)
}
