% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngl.R
\name{download_estimated_velocities_ngl}
\alias{download_estimated_velocities_ngl}
\title{Download estimated velocities using the MIDAS estimator provided by the Nevada Geodetic Laboratory (NGL)}
\usage{
download_estimated_velocities_ngl()
}
\value{
Return a \code{data.frame} with all stations name, information about the time series for each station, estimated velocities and estimated standard deviation of the estimated velocities.
}
\description{
Download estimated velocities using the MIDAS estimator provided by the Nevada Geodetic Laboratory (NGL)
}
