% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{inv_trans_alpha_matern}
\alias{inv_trans_alpha_matern}
\title{Inverse transform for Matérn alpha}
\usage{
inv_trans_alpha_matern(x)
}
\arguments{
\item{x}{Numeric vector with values > 1/2.}
}
\value{
Numeric vector on the real line.
}
\description{
Maps domain values (> 1/2) back to the real line.
}
\keyword{internal}
