% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ngl.R
\name{load_station_ngl}
\alias{load_station_ngl}
\title{Load station data from Nevada Geodetic Laboratory}
\usage{
load_station_ngl(station_name)
}
\description{
Load station data from Nevada Geodetic Laboratory
}
