% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{matern}
\alias{matern}
\title{Matern process}
\usage{
matern(sigma2 = NULL, lambda = NULL, alpha = NULL)
}
\arguments{
\item{sigma2}{Marginal variance (> 0).}

\item{lambda}{Range/scale parameter (> 0).}

\item{alpha}{Smoothness parameter in (1/2, 10).}
}
\value{
A \code{time_series_model} object.
}
\description{
Constructs a \code{time_series_model} for a Matern covariance process with
variance \code{sigma2}, range \code{lambda}, and smoothness \code{alpha}.
The autocovariance is
\eqn{ \gamma(h) = \mathrm{cov}(X_t, X_{t+h}) = \frac{2 \sigma^2}{\Gamma(\alpha-1 / 2) 2^{\alpha-1 / 2}}|\lambda h|^{\alpha-1 / 2} \mathcal{K}_{|\alpha-1 / 2|}(| \lambda h|)}
where \eqn{\mathcal{K}_\omega(x)} is the modified Bessel function of the second kind of order \eqn{\omega}.
}
\examples{
mod <- matern(sigma2 = 1, lambda = 0.2, alpha = 1.0)
mod
}
\references{
Lilly JM, Sykulski AM, Early JJ, Olhede SC (2017). "Fractional Brownian motion,
the Matérn process, and stochastic modeling of turbulent dispersion."
\emph{Nonlinear Processes in Geophysics}, 24(3), 481-514.
}
