% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{pl}
\alias{pl}
\title{Stationary Power-Law process (\code{time_series_model})}
\usage{
pl(kappa = NULL, sigma2 = NULL)
}
\arguments{
\item{kappa}{Power-law parameter in (-1, 1).}

\item{sigma2}{Process variance (> 0).}
}
\value{
A \code{time_series_model} object.
}
\description{
Constructs a \code{time_series_model} representing a stationary power-law
process with parameters \code{kappa} and \code{sigma2}.
In the frequency domain, a power-law process is often described by a
spectrum \eqn{P(f) = P_0 f^{\kappa}} (Bos et al., 2008), where \eqn{f} is the frequency, \eqn{P_0} is a constant and \eqn{\kappa} is the spectral index.
Note that we use the convention that the power spectral density satisfies
\eqn{P(f) \propto |f|^{\kappa}}, where \eqn{\kappa > -1} ensures second-order
stationarity. This corresponds to the alternative notation
\eqn{P(f) \propto |f|^{-\alpha}} with \eqn{\alpha = -\kappa}.
The autocovariance \eqn{\gamma(h) = \mathrm{cov}(X_t, X_{t+h})} used here (Hosking, 1981) is
\eqn{\gamma(0) = \sigma^{2} \frac{\Gamma(1+\kappa)}{\Gamma\left(1+\kappa/2\right)^2}},
and for \eqn{h > 0}{h > 0}
\eqn{\gamma(h) = \frac{-\kappa/2 + h - 1}{\kappa/2 + h}\,\gamma(h-1)}.
}
\examples{
mod <- pl(kappa = -0.5, sigma2 = 2)
mod
}
\references{
Bos MS, Fernandes RMS, Williams SDP, Bastos L (2008). "Fast error analysis
of continuous GPS observations." \emph{Journal of Geodesy}, 82, 157-166.

Hosking JRM (1981). "Fractional differencing." \emph{Biometrika}, 68(1), 165-176.
}
