% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwmx2.R
\name{plot.fit_gnss_ts_ngl}
\alias{plot.fit_gnss_ts_ngl}
\title{Plot a \code{fit_gnss_ts_ngl} object}
\usage{
\method{plot}{fit_gnss_ts_ngl}(x, ...)
}
\arguments{
\item{x}{A \code{fit_gnss_ts_ngl} object.}

\item{...}{Additional graphical parameters.}
}
\value{
No return value. Plot a \code{fit_gnss_ts_ngl} object.
}
\description{
Plot a \code{fit_gnss_ts_ngl} object
}
