% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{plot.generated_composite_model_time_series}
\alias{plot.generated_composite_model_time_series}
\title{Plot a \code{generated_composite_model_time_series} object}
\usage{
\method{plot}{generated_composite_model_time_series}(x, ...)
}
\arguments{
\item{x}{A \code{generated_composite_model_time_series}.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces stacked line plots for each component and the sum for a
\code{generated_composite_model_time_series} object.
}
\examples{
m2 <- wn(sigma2 = 1) + ar1(phi = 0.8, sigma2 = 0.5)
y2 <- generate(m2, n = 200, seed = 123)
plot(y2)
}
