% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{plot.generated_missingness}
\alias{plot.generated_missingness}
\title{Plot a generated missingness process}
\usage{
\method{plot}{generated_missingness}(x, ...)
}
\arguments{
\item{x}{A \code{generated_missingness}.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a step plot for a \code{generated_missingness} object.
}
\examples{
m0 <- markov_two_states()(p1 = 0.05, p2 = 0.9)
z0 <- generate(m0, n = 200, seed = 123)
plot(z0)
}
