% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_model.R
\name{plot.generated_time_series}
\alias{plot.generated_time_series}
\title{Plot a \code{generated_time_series} object}
\usage{
\method{plot}{generated_time_series}(x, ...)
}
\arguments{
\item{x}{A \code{generated_time_series}.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a single line plot for a \code{generated_time_series} object.
}
\examples{
m1 <- wn(sigma2 = 1)
y1 <- generate(m1, n = 200, seed = 123)
plot(y1)
}
