% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwm2.R
\name{plot.gmwm2_fit}
\alias{plot.gmwm2_fit}
\title{Plot method for a \code{gmwm2_fit} object}
\usage{
\method{plot}{gmwm2_fit}(
  x,
  show_ci = TRUE,
  col_emp = "black",
  col_theo = "darkorange",
  col_ci = "#e6f7fb",
  lwd = 2,
  pch_emp = 16,
  pch_theo = 21,
  cex_theo = 1.4,
  legend_pos = "auto",
  ...
)
}
\arguments{
\item{x}{A \code{gmwm2_fit} object.}

\item{show_ci}{Logical; if TRUE and available, show empirical CI bars.}

\item{col_emp}{Color for empirical WV points/line.}

\item{col_theo}{Color for theoretical WV line.}

\item{col_ci}{Color for empirical WV CI band.}

\item{lwd}{Line width for theoretical curve.}

\item{pch_emp}{Plotting character for empirical points.}

\item{pch_theo}{Plotting character for theoretical points.}

\item{cex_theo}{Size for theoretical points.}

\item{legend_pos}{Legend position (e.g., "topleft") or "auto".}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
The input object, invisibly.
}
\description{
Plots empirical wavelet variance with the fitted theoretical curve and,
for sum models, component-implied theoretical curves.
}
\examples{
n = 10000
mod = wn(20) + ar1(phi = .995, sigma2 = .2)
y = generate(mod, n = n, seed = 123)
plot(y)
fit = gmwm2(y, model = wn() + ar1() )
fit
plot(fit)
}
