% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_time_series_model.R
\name{+.time_series_model}
\alias{+.time_series_model}
\title{Add to a \code{time_series_model} object}
\usage{
\method{+}{time_series_model}(e1, e2)
}
\arguments{
\item{e1}{Left operand.}

\item{e2}{Right operand.}
}
\value{
A \code{sum_model}.
}
\description{
Combines \code{time_series_model} and/or \code{sum_model} into a \code{sum_model}.
}
\examples{
m1 <- wn(sigma2 = 1)
m2 <- ar1(phi = 0.8, sigma2 = 0.5)
model <- m1 + m2
model
}
