% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwm2.R
\name{print.gmwm2_fit}
\alias{print.gmwm2_fit}
\title{Print method for gmwm2_fit}
\usage{
\method{print}{gmwm2_fit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{gmwm2_fit} object.}

\item{digits}{Significant digits for printing.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print method for gmwm2_fit
}
