% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_time_series_model.R
\name{sum_model}
\alias{sum_model}
\title{Sum of stochastic models (internal)}
\usage{
sum_model(models)
}
\arguments{
\item{models}{A list of \code{time_series_model} objects.}
}
\value{
A \code{sum_model} object.
}
\description{
Creates a composite model representing a sum of independent
\code{time_series_model} components. This is an internal constructor used
by the \code{+} methods.
}
\examples{
mod <- pl(kappa = 0.5, sigma2 = 2) + wn(sigma2 = 1)
print(mod)
}
\keyword{internal}
