% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwmx2.R
\name{summary.fit_gnss_ts_ngl}
\alias{summary.fit_gnss_ts_ngl}
\title{Extract estimated parameters from a \code{fit_gnss_ts_ngl}}
\usage{
\method{summary}{fit_gnss_ts_ngl}(object, ...)
}
\arguments{
\item{object}{A \code{fit_gnss_ts_ngl} object.}

\item{...}{Additional parameters.}
}
\description{
Extract estimated parameters from a \code{fit_gnss_ts_ngl}
}
\examples{
x = download_station_ngl("CHML")
fit = gmwmx2(x,n_seasonal = 2, component = "N")
summary(fit)
}
