% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngl.R
\name{download_all_stations_ngl}
\alias{download_all_stations_ngl}
\title{Download all stations name and location from the Nevada Geodetic Laboratory}
\usage{
download_all_stations_ngl(verbose = FALSE)
}
\arguments{
\item{verbose}{A \code{boolean} that controls the level of detail in the output of the \code{wget} command used to load data. Default is \code{FALSE}.}
}
\value{
Return a \code{data.frame} with all stations name, latitude, longitude and heights.
}
\description{
Download all stations name and location from the Nevada Geodetic Laboratory
}
\examples{
df_all_stations <- download_all_stations_ngl()
head(df_all_stations)

}
