% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngl.R
\name{download_estimated_velocities_ngl}
\alias{download_estimated_velocities_ngl}
\title{Download estimated velocities provided by the Nevada Geodetic Laboratory for all stations.}
\usage{
download_estimated_velocities_ngl(verbose = FALSE)
}
\arguments{
\item{verbose}{A \code{boolean} that controls the level of detail in the output of the \code{wget} command used to load data. Default is \code{FALSE}.}
}
\value{
Return a \code{data.frame} with all stations name, information about the time series for each station, estimated velocities and estimated standard deviation of the estimated velocities.
}
\description{
Download estimated velocities provided by the Nevada Geodetic Laboratory for all stations.
}
\examples{
df_estimated_velocities <- download_estimated_velocities_ngl()
head(df_estimated_velocities)
}
