% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwmx2.R
\name{gmwmx2}
\alias{gmwmx2}
\title{Estimate a linear model with White noise and Flicker noise for the residuals in presence of missing data using the GMWMX estimator.}
\usage{
gmwmx2(
  x,
  n_seasonal = 2,
  vec_earthquakes_relaxation_time = NULL,
  component = "N"
)
}
\arguments{
\item{x}{A \code{gnss_ts_ngl} object.}

\item{n_seasonal}{An \code{integer} specifying the number of seasonal signals in the time series.}

\item{vec_earthquakes_relaxation_time}{A \code{vecor} specifying the relaxation time for each earthquakes indicated for the time series.}

\item{component}{A \code{string} with value either "N", "E" or "V" that specify which component to estimate (Northing, Easting or Vertical).}
}
\description{
Estimate a linear model with White noise and Flicker noise for the residuals in presence of missing data using the GMWMX estimator.
}
