% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{gmwmx2_plot_colors}
\alias{gmwmx2_plot_colors}
\title{Plot colors for generated time series}
\format{
A character vector of hex colors.
}
\usage{
gmwmx2_plot_colors
}
\description{
Editable default palette used by the plotting methods for generated time
series. When more colors are needed, the palette is expanded with
interpolation.
}
\seealso{
.gmwmx2_get_plot_colors
}
\keyword{internal}
