% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwm2.R
\name{plot.gmwm2_fit}
\alias{plot.gmwm2_fit}
\title{Plot method for a \code{gmwm2_fit} object}
\usage{
\method{plot}{gmwm2_fit}(
  x,
  show_ci = TRUE,
  col_emp = "black",
  col_theo = "darkorange",
  col_ci = "#e6f7fb",
  lwd = 2,
  pch_emp = 16,
  pch_theo = 21,
  cex_theo = 1.4,
  legend_pos = "auto",
  ...
)
}
\arguments{
\item{x}{A \code{gmwm2_fit} object.}

\item{show_ci}{Logical; if TRUE and available, show empirical CI bars.}

\item{col_emp}{Color for empirical WV points/line.}

\item{col_theo}{Color for theoretical WV line.}

\item{lwd}{Line width for theoretical curve.}

\item{...}{Additional arguments passed to \code{plot()}.}

\item{pch}{Plotting character for empirical points.}
}
\value{
The input object, invisibly.
}
\description{
Plots empirical wavelet variance with the fitted theoretical curve.
}
\examples{
n = 1000
mod4 = wn(20) + rw(.1)
y4 = generate(mod4, n = n)
plot(y4)
fit4 = gmwm2(y4, model = wn() + rw())
fit4
plot(fit4)
}
