% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwmx2.R
\name{plot.gmwmx2_fit_gnss_ts_ngl}
\alias{plot.gmwmx2_fit_gnss_ts_ngl}
\title{Plot a \code{gmwmx2_fit_gnss_ts_ngl} object}
\usage{
\method{plot}{gmwmx2_fit_gnss_ts_ngl}(x, ...)
}
\arguments{
\item{x}{A \code{gmwmx2_fit_gnss_ts_ngl} object.}

\item{...}{Additional graphical parameters.}
}
\value{
No return value. Plot a \code{gmwmx2_fit_gnss_ts_ngl} object.
}
\description{
Plot a \code{gmwmx2_fit_gnss_ts_ngl} object
}
\examples{
# station_data = gmwmx2::download_station_ngl("1LSU")
# fit station with WN and PL
# fit1 <- gmwmx2(
#   station_data,
#   n_seasonal = 2,
#   model = wn() + pl(), component = "N"
# )
# fit1
# plot(fit1)

}
