% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwmx2_new.R
\name{plot.gmwmx2_fit_gnss_ts_ngl}
\alias{plot.gmwmx2_fit_gnss_ts_ngl}
\title{Plot a \code{gmwmx2_fit_gnss_ts_ngl} object}
\usage{
\method{plot}{gmwmx2_fit_gnss_ts_ngl}(x, ...)
}
\arguments{
\item{x}{A \code{gmwmx2_fit_gnss_ts_ngl} object.}

\item{...}{Additional graphical parameters.}
}
\value{
No return value. Plot a \code{gmwmx2_fit_gnss_ts_ngl} object.
}
\description{
Plot a \code{gmwmx2_fit_gnss_ts_ngl} object
}
\examples{
fit station with WN and AR1
fit1 <- gmwmx2_new(
  station_data,
  n_seasonal = 2,
  model = wn() + ar1(), component = "N"
)
fit1
plot(fit1)
fit station with WN and PL
fit2 <- gmwmx2_new(
  station_data,
  n_seasonal = 2,
  model = wn() + pl(), component = "N"
)
fit2
plot(fit2)
fit station with WN and Matern
fit3 = gmwmx2_new(
  station_data,
  n_seasonal = 2,
  model = wn() + matern(), component = "N"
)
fit3
plot(fit3)

}
