% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngl.R
\name{plot.gnss_ts_ngl}
\alias{plot.gnss_ts_ngl}
\title{Plot a \code{gnss_ts_ngl} object}
\usage{
\method{plot}{gnss_ts_ngl}(x, component = NULL, ...)
}
\arguments{
\item{x}{A \code{gnss_ts_ngl} object.}

\item{component}{A \code{string} with value either "N", "E" or "V" that specify which component to plot (Northing, Easting or Vertical).}

\item{...}{Additional graphical parameters.}
}
\value{
No return value. Plot a \code{gnss_ts_ngl} object.
}
\description{
Plot a \code{gnss_ts_ngl} object
}
\examples{
station_1LSU <- download_station_ngl("1LSU")
plot(station_1LSU)
plot(station_1LSU, component = "N")
plot(station_1LSU, component = "E")
plot(station_1LSU, component = "V")
}
