% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_time_series_model.R
\name{+.sum_model}
\alias{+.sum_model}
\title{Add to a \code{sum_model} object}
\usage{
\method{+}{sum_model}(e1, e2)
}
\arguments{
\item{e1}{Left operand.}

\item{e2}{Right operand.}
}
\value{
A \code{sum_model}.
}
\description{
Add to a \code{sum_model} object
}
\examples{
m1 <- wn(sigma2 = 1)
m2 <- ar1(phi = 0.8, sigma2 = 0.5)
m3 <- pl(kappa = 0.3, sigma2 = 2)
model <- (m1 + m2) + m3
}
