% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwm2.R
\name{print.gmwm2_fit}
\alias{print.gmwm2_fit}
\title{Print method for a \code{gmwm2_fit} object}
\usage{
\method{print}{gmwm2_fit}(x, digits = 4, show_initial_parameters = FALSE, ...)
}
\arguments{
\item{x}{A \code{gmwm2_fit} object.}

\item{digits}{Significant digits for printing.}

\item{show_initial_parameters}{Logical; if TRUE, also show the initial parameters used for optimization.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print method for a \code{gmwm2_fit} object
}
\examples{
n = 10000
mod = wn(20) + ar1(phi = .995, sigma2 = .2)
y = generate(mod, n = n, seed = 123)
plot(y)
fit = gmwm2(y, model = wn() + ar1() )
fit
}
